Moved `convert_mapping` from `ged.py` to `bipartite_ged.py`
